Fix some warnings for MS Windows compiler

+++ Source/Misc/base64.cpp
@@ -60,7 +60,7 @@ namespace INMOST
 	
 	std::vector<BYTE> base64_decode(std::string const& encoded_string)
 	{
-		int in_len = encoded_string.size();
+        int in_len = static_cast<int>(encoded_string.size());
 		int i = 0;
 		int j = 0;
 		int in_ = 0;
@@ -73,7 +73,7 @@ namespace INMOST
 			if (i ==4)
 			{
 				for (i = 0; i <4; i++)
-					char_array_4[i] = base64_chars.find(char_array_4[i]);
+                    char_array_4[i] = (BYTE)base64_chars.find(char_array_4[i]);
 		  
 				char_array_3[0] = (char_array_4[0] << 2) + ((char_array_4[1] & 0x30) >> 4);
 				char_array_3[1] = ((char_array_4[1] & 0xf) << 4) + ((char_array_4[2] & 0x3c) >> 2);
@@ -91,7 +91,7 @@ namespace INMOST
 				char_array_4[j] = 0;
 	  
 			for (j = 0; j <4; j++)
-				char_array_4[j] = base64_chars.find(char_array_4[j]);
+                char_array_4[j] = (BYTE)base64_chars.find(char_array_4[j]);
 	  
 			char_array_3[0] = (char_array_4[0] << 2) + ((char_array_4[1] & 0x30) >> 4);
 			char_array_3[1] = ((char_array_4[1] & 0xf) << 4) + ((char_array_4[2] & 0x3c) >> 2);
